Move emeter related functions to tapodevice
